introduce DIR_FTP_INCOMING constant
git-svn-id: http://argo.unfuddle.com/svn/argo_screenwerk/trunk@126 7a690eda-67ec-4556-b061-4c86d960b817

+++ www/application/controllers/screen.php
@@ -83,6 +83,7 @@ class Screen extends Controller {
          $_source = DIR_FTP_MASTERS . '/' . $media['id'] . '.' . $media['type'];
          $_destination = DIR_FTP_SCREENS . '/' . $screen_id . '/' . $media['id'] . '.' . $media['type'];
 
+         if(!is_file($_source)) die($_source);
          if(is_file($_destination)) unlink($_destination);
          link($_source, $_destination);
       }
